Move ViewportScrollMix code to Viewport component
