+++ packages/react-data-grid/src/Viewport.js
@@ -1,9 +1,16 @@
 const React                = require('react');
+const ReactDOM             = require('react-dom');
 const createReactClass = require('create-react-class');
 const Canvas               = require('./Canvas');
 const ViewportScroll       = require('./ViewportScrollMixin');
 const cellMetaDataShape    = require('./PropTypeShapes/CellMetaDataShape');
 import PropTypes from 'prop-types';
+import ColumnUtils from './ColumnUtils';
+import {
+  getGridState,
+  getRenderedColumnCount,
+  getNextScrollState
+} from './utils/viewportUtils';
 
 const Viewport = createReactClass({
   displayName: 'Viewport',
@@ -64,6 +71,107 @@ const Viewport = createReactClass({
 
   setScrollLeft(scrollLeft: number) {
     this.canvas.setScrollLeft(scrollLeft);
+  },
+
+  getDefaultProps(): { rowHeight: number } {
+    return {
+      rowHeight: 30
+    };
+  },
+
+  getInitialState() {
+    return getGridState(this.props);
+  },
+
+  getDOMNodeOffsetWidth() {
+    return ReactDOM.findDOMNode(this).offsetWidth;
+  },
+
+  getRenderedColumnCount(displayStart, width) {
+    return getRenderedColumnCount(this.props, this.getDOMNodeOffsetWidth, displayStart, width);
+  },
+
+  clearScrollTimer() {
+    if (this.resetScrollStateTimeoutId) {
+      clearTimeout(this.resetScrollStateTimeoutId);
+    }
+  },
+
+  resetScrollStateAfterDelay() {
+    this.clearScrollTimer();
+    this.resetScrollStateTimeoutId = setTimeout(
+      this.resetScrollStateAfterDelayCallback,
+      500
+    );
+  },
+
+  resetScrollStateAfterDelayCallback() {
+    this.resetScrollStateTimeoutId = null;
+    this.setState({
+      isScrolling: false
+    });
+  },
+
+  updateScroll(scrollTop: number, scrollLeft: number, height: number, rowHeight: number, length: number, width) {
+    this.resetScrollStateAfterDelay();
+    const nextScrollState = getNextScrollState(this.props, this.getDOMNodeOffsetWidth, scrollTop, scrollLeft, height, rowHeight, length, width);
+
+    this.setState(nextScrollState);
+  },
+
+  metricsUpdated() {
+    let height = this.DOMMetrics.viewportHeight();
+    let width = this.DOMMetrics.viewportWidth();
+    if (height) {
+      this.updateScroll(
+        this.state.scrollTop,
+        this.state.scrollLeft,
+        height,
+        this.props.rowHeight,
+        this.props.rowsCount,
+        width
+      );
+    }
+  },
+
+  componentWillReceiveProps(nextProps: { rowHeight: number; rowsCount: number, rowOffsetHeight: number }) {
+    if (this.props.rowHeight !== nextProps.rowHeight ||
+      this.props.minHeight !== nextProps.minHeight) {
+      const newState = getGridState(nextProps);
+      this.updateScroll(
+        newState.scrollTop,
+        newState.scrollLeft,
+        newState.height,
+        nextProps.rowHeight,
+        nextProps.rowsCount
+      );
+    } else if (ColumnUtils.getSize(this.props.columnMetrics.columns) !== ColumnUtils.getSize(nextProps.columnMetrics.columns)) {
+      this.setState(getGridState(nextProps));
+    } else if (this.props.rowsCount !== nextProps.rowsCount) {
+      this.updateScroll(
+        this.state.scrollTop,
+        this.state.scrollLeft,
+        this.state.height,
+        nextProps.rowHeight,
+        nextProps.rowsCount
+      );
+      // Added to fix the hiding of the bottom scrollbar when showing the filters.
+    } else if (this.props.rowOffsetHeight !== nextProps.rowOffsetHeight) {
+      // The value of height can be positive or negative and will be added to the current height to cater for changes in the header height (due to the filer)
+      let height = this.props.rowOffsetHeight - nextProps.rowOffsetHeight;
+
+      this.updateScroll(
+        this.state.scrollTop,
+        this.state.scrollLeft,
+        this.state.height + height,
+        nextProps.rowHeight,
+        nextProps.rowsCount
+      );
+    }
+  },
+
+  componentWillUnmount() {
+    this.clearScrollTimer();
   },
 
   render() {
